update [ergo]: general cleanup
