+++ src/chains/ergo/ergo.ts
@@ -3,8 +3,6 @@ import {
   SecretKey,
   SecretKeys,
   Wallet,
-  ErgoBoxes,
-  UnsignedTransaction,
   Mnemonic,
   ExtSecretKey,
   DerivationPath,
@@ -25,49 +23,39 @@ import {
   AmmPool,
   makeNativePools,
   makeWrappedNativePoolActionsSelector,
-  minValueForOrder,
-  minValueForSetup,
   SwapExtremums,
   SwapParams,
   swapVars,
 } from '@patternglobal/ergo-dex-sdk';
 import {
   Explorer,
-  Prover,
   ErgoTx,
-  UnsignedErgoTx,
-  unsignedErgoTxToProxy,
   DefaultTxAssembler,
   AssetAmount,
-  MinBoxValue,
-  DefaultBoxSelector,
-  InsufficientInputs,
   publicKeyFromAddress,
   TransactionContext,
-  Address,
-  BoxSelection,
-  Input as TxInput,
   RustModule,
 } from '@patternglobal/ergo-sdk';
-import { makeTarget } from '@patternglobal/ergo-dex-sdk/build/main/utils/makeTarget';
 import { NativeExFeeType } from '@patternglobal/ergo-dex-sdk/build/main/types';
 import { NetworkContext } from '@patternglobal/ergo-sdk/build/main/entities/networkContext';
 import { ErgoNetwork } from './types/ergo.type';
+import { getBaseInputParameters, getInputs, getTxContext } from './ergo.util';
+import { WalletProver } from './wallet-prover.service';
 
 class Pool extends AmmPool {
-  private name: string;
+  private _name: string;
 
   constructor(public pool: AmmPool) {
     super(pool.id, pool.lp, pool.x, pool.y, pool.poolFeeNum);
 
-    this.name = `${this.x.asset.name}/${this.y.asset.name}`;
+    this._name = `${this.x.asset.name}/${this.y.asset.name}`;
   }
 
-  public get getName() {
-    return this.name;
+  public get name() {
+    return this._name;
   }
 
-  public get getPoolInfo() {
+  public get info() {
     return {
       id: this.id,
       lp: this.lp,
@@ -76,136 +64,6 @@ class Pool extends AmmPool {
       feeNum: this.feeNum,
       feeDenom: this.feeDenom,
     };
-  }
-
-  // calculatePriceImpact(input: any): number {
-  //   const ratio =
-  //     input.asset.id === this.x.asset.id
-  //       ? math.evaluate!(
-  //         `${renderFractions(this.y.amount.valueOf(), this.y.asset.decimals)} / ${renderFractions(this.x.amount.valueOf(), this.x.asset.decimals)}`,
-  //       ).toString()
-  //       : math.evaluate!(
-  //         `${renderFractions(this.x.amount.valueOf(), this.x.asset.decimals)} / ${renderFractions(this.y.amount.valueOf(), this.y.asset.decimals)}`,
-  //       ).toString();
-  //   const outputAmount = calculatePureOutputAmount(input, this);
-  //   const outputRatio = math.evaluate!(
-  //     `${outputAmount} / ${renderFractions(input.amount, input.asset.decimals)}`,
-  //   ).toString();
-  //
-  //   return Math.abs(
-  //     math.evaluate!(`(${outputRatio} * 100 / ${ratio}) - 100`).toFixed(2),
-  //   );
-  // }
-}
-
-export type BaseInputParameters = {
-  baseInput: AssetAmount;
-  baseInputAmount: bigint;
-  minOutput: AssetAmount;
-};
-export const getBaseInputParameters = (
-  pool: AmmPool,
-  { inputAmount, slippage }: { inputAmount: any; slippage: number },
-): BaseInputParameters => {
-  const baseInputAmount =
-    inputAmount.asset.id === pool.x.asset.id
-      ? pool.x.withAmount(inputAmount.amount)
-      : pool.y.withAmount(inputAmount.amount);
-  const minOutput = pool.outputAmount(baseInputAmount as any, slippage);
-
-  return {
-    baseInput: baseInputAmount as any,
-    baseInputAmount: inputAmount.amount,
-    minOutput: minOutput as any,
-  };
-};
-export const getInputs = (
-  utxos: ErgoBox[],
-  assets: AssetAmount[],
-  fees: { minerFee: bigint; uiFee: bigint; exFee: bigint },
-  minBoxValue: bigint,
-  ignoreMinBoxValue?: boolean,
-  setup?: boolean,
-): BoxSelection => {
-  let minFeeForOrder = minValueForOrder(fees.minerFee, fees.uiFee, fees.exFee);
-  if (setup) {
-    minFeeForOrder = minValueForSetup(fees.minerFee, fees.uiFee);
-  }
-  if (ignoreMinBoxValue) {
-    minFeeForOrder -= MinBoxValue;
-  }
-
-  const target = makeTarget(assets, minFeeForOrder);
-
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-expect-error
-  const inputs = DefaultBoxSelector.select(utxos, target, minBoxValue);
-
-  if (inputs instanceof InsufficientInputs) {
-    throw new Error(
-      `Error in getInputs function: InsufficientInputs -> ${inputs}`,
-    );
-  }
-
-  return inputs;
-};
-export const getTxContext = (
-  inputs: BoxSelection,
-  network: NetworkContext,
-  address: Address,
-  minerFee: bigint,
-): TransactionContext => ({
-  inputs,
-  selfAddress: address,
-  changeAddress: address,
-  feeNErgs: minerFee,
-  network,
-});
-
-export class WalletProver implements Prover {
-  readonly wallet: Wallet;
-  readonly nodeService: NodeService;
-
-  constructor(wallet: Wallet, nodeService: NodeService) {
-    this.wallet = wallet;
-    this.nodeService = nodeService;
-  }
-
-  /** Sign the given transaction.
-   */
-  async sign(tx: UnsignedErgoTx): Promise<ErgoTx> {
-    const ctx = await this.nodeService.getCtx();
-    const proxy = unsignedErgoTxToProxy(tx);
-    const wasmtx = UnsignedTransaction.from_json(JSON.stringify(proxy));
-    try {
-      return this.wallet
-        .sign_transaction(
-          ctx,
-          wasmtx,
-          ErgoBoxes.from_boxes_json(proxy.inputs),
-          ErgoBoxes.empty(),
-        )
-        .to_js_eip12();
-    } catch {
-      throw new Error('not be able to sign!');
-    }
-  }
-
-  async submit(tx: ErgoTx): Promise<ErgoTx> {
-    const txId = await this.nodeService.postTransaction(tx);
-    return {
-      ...tx,
-      id: txId,
-    };
-  }
-
-  // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-  // @ts-expect-error
-  // eslint-disable-next-line @typescript-eslint/no-unused-vars
-  signInput(tx: UnsignedErgoTx, input: number): Promise<TxInput> {
-    // eslint-disable-next-line @typescript-eslint/ban-ts-comment
-    // @ts-expect-error
-    return;
   }
 }
 
@@ -517,6 +375,7 @@ export class Ergo {
       BigInt(0),
     );
     const assets: Record<string, bigint> = {};
+
     utxos.forEach((box) => {
       box.assets.forEach((asset) => {
         if (Object.keys(assets).includes(asset.tokenId))
@@ -524,6 +383,7 @@ export class Ergo {
         else assets[asset.tokenId] = BigInt(asset.amount);
       });
     });
+
     return { balance, assets };
   }
 
@@ -547,6 +407,7 @@ export class Ergo {
   private async loadPools(): Promise<void> {
     let offset = 0;
     let pools: Array<Pool> = await this.getPoolData(this.poolLimit, offset);
+
     while (pools.length > 0) {
       for (const pool of pools) {
         if (!this.ammPools.filter((ammPool) => ammPool.id === pool.id).length) {
@@ -562,12 +423,14 @@ export class Ergo {
   public async loadPool(poolId: string): Promise<void> {
     await RustModule.load(true);
     const pool = await this.getPool(poolId);
-    if (!pool)
+
+    if (!pool) {
       this.ammPools.push(
         // eslint-disable-next-line @typescript-eslint/ban-ts-comment
         // @ts-expect-error
         new Pool(await makeNativePools(this._explorer).get(poolId)),
       );
+    }
   }
 
   private async getPoolData(limit: number, offset: number): Promise<any> {
@@ -575,6 +438,7 @@ export class Ergo {
       limit,
       offset,
     });
+
     return AmmPool;
   }
 
@@ -685,8 +549,8 @@ export class Ergo {
       return_address,
       config.network.defaultMinerFee,
     );
-
     const actions = poolActions(pool);
+
     return await actions.swap(swapParams, txContext);
   }
 
@@ -761,6 +625,7 @@ export class Ergo {
       inputAmount: from,
       slippage: slippage || config.network.defaultSlippage,
     });
+
     return minOutput;
   }
